+++ lumigator/frontend/src/App.vue
@@ -60,10 +60,10 @@
 
 <script lang="ts" setup>
 import { onMounted, ref } from 'vue'
-import LMenu from '@/components/organisms/LMenu.vue'
+import LMenu from '@/components/layout/LMenu.vue'
 import { useDatasetStore } from '@/stores/datasetsStore'
 import { useExperimentStore } from '@/stores/experimentsStore'
-import { useSlidePanel } from '@/composables/SlidingPanel'
+import { useSlidePanel } from '@/composables/useSlidePanel'
 import ConfirmDialog from 'primevue/confirmdialog'
 import Toast from 'primevue/toast'
 import Button from 'primevue/button'
