+++ lumigator/frontend/src/components/datasets/LDatasetTable.vue
@@ -67,7 +67,7 @@ import { ref, computed, onMounted, watch } from 'vue'
 import DataTable from 'primevue/datatable'
 import Column from 'primevue/column'
 import Menu from 'primevue/menu'
-import { useSlidePanel } from '@/composables/SlidingPanel'
+import { useSlidePanel } from '@/composables/useSlidePanel'
 import type { MenuItem } from 'primevue/menuitem'
 import type { Dataset } from '@/types/Dataset'
 import { formatDate } from '@/helpers/formatDate'
