+++ lumigator/frontend/src/components/experiments/LExperimentForm.vue
@@ -92,7 +92,7 @@ import Select from 'primevue/select'
 import Textarea from 'primevue/textarea'
 import InputText from 'primevue/inputtext'
 import InputNumber from 'primevue/inputnumber'
-import LModelCards from '@/components/molecules/LModelCards.vue'
+import LModelCards from '@/components/experiments/LModelCards.vue'
 import { useToast } from 'primevue/usetoast'
 import type { ToastMessageOptions } from 'primevue'
 import type { Dataset } from '@/types/Dataset'
