+++ lumigator/frontend/src/components/experiments/LExperimentResults.vue
@@ -118,7 +118,7 @@ import { computed, ref, onUnmounted } from 'vue'
 import { useExperimentStore } from '@/stores/experimentsStore'
 import { useModelStore } from '@/stores/modelsStore'
 import { storeToRefs } from 'pinia'
-import LJobResults from '@/components/molecules/LJobResults.vue'
+import LJobResults from '@/components/experiments/LJobResults.vue'
 import DataTable from 'primevue/datatable'
 import Column from 'primevue/column'
 import type { Model } from '@/types/Experiment'
