+++ lumigator/frontend/src/components/experiments/LExperimentTable.vue
@@ -87,9 +87,9 @@ import { storeToRefs } from 'pinia'
 import DataTable, { type DataTableRowClickEvent } from 'primevue/datatable'
 import Column from 'primevue/column'
 
-import { useSlidePanel } from '@/composables/SlidingPanel'
+import { useSlidePanel } from '@/composables/useSlidePanel'
 import Tag from 'primevue/tag'
-import LJobsTable from '@/components/molecules/LJobsTable.vue'
+import LJobsTable from '@/components/jobs/LJobsTable.vue'
 import { useExperimentStore } from '@/stores/experimentsStore'
 import type { Experiment, Job } from '@/types/Experiment'
 import { formatDate } from '@/helpers/formatDate'
