+++ lumigator/frontend/src/components/experiments/LExperimentsEmpty.vue
