+++ lumigator/frontend/src/components/health/LHealthStatus.vue
