+++ lumigator/frontend/src/components/views/LDatasets.vue
@@ -76,17 +76,17 @@ import { onMounted, ref } from 'vue'
 import { storeToRefs } from 'pinia'
 import { useDatasetStore } from '@/stores/datasetsStore'
 import { useExperimentStore } from '@/stores/experimentsStore'
-import { useSlidePanel } from '@/composables/SlidingPanel'
+import { useSlidePanel } from '@/composables/useSlidePanel'
 import { useRouter } from 'vue-router'
-import LPageHeader from '@/components/molecules/LPageHeader.vue'
-import LDatasetTable from '@/components/molecules/LDatasetTable.vue'
-import LFileUpload from '@/components/molecules/LFileUpload.vue'
-import LDatasetEmpty from '@/components/molecules/LDatasetEmpty.vue'
-import LDatasetDetails from '@/components/organisms/LDatasetDetails.vue'
-import LInferenceJobsTable from '@/components/molecules/LInferenceJobsTable.vue'
-import LExperimentDetails from '@/components/molecules/LExperimentDetails.vue'
-import LExperimentsDrawer from '@/components/molecules/LExperimentsDrawer.vue'
-import LExperimentLogs from '@/components/molecules/LExperimentLogs.vue'
+import LPageHeader from '@/components/layout/LPageHeader.vue'
+import LDatasetTable from '@/components/datasets/LDatasetTable.vue'
+import LFileUpload from '@/components/common/LFileUpload.vue'
+import LDatasetEmpty from '@/components/datasets/LDatasetEmpty.vue'
+import LDatasetDetails from '@/components/datasets/LDatasetDetails.vue'
+import LInferenceJobsTable from '@/components/jobs/LInferenceJobsTable.vue'
+import LExperimentDetails from '@/components/experiments/LExperimentDetails.vue'
+import LExperimentsDrawer from '@/components/experiments/LExperimentsDrawer.vue'
+import LExperimentLogs from '@/components/experiments/LExperimentLogs.vue'
 import Tabs from 'primevue/tabs'
 import TabList from 'primevue/tablist'
 import Tab from 'primevue/tab'
