+++ lumigator/frontend/src/components/views/LExperiments.vue
@@ -55,16 +55,16 @@ import { storeToRefs } from 'pinia'
 import { useExperimentStore } from '@/stores/experimentsStore'
 import { useDatasetStore } from '@/stores/datasetsStore'
 import { useModelStore } from '@/stores/modelsStore'
-import { useSlidePanel } from '@/composables/SlidingPanel'
-import LPageHeader from '@/components/molecules/LPageHeader.vue'
-import LExperimentTable from '@/components/molecules/LExperimentTable.vue'
-import LExperimentForm from '@/components/molecules/LExperimentForm.vue'
-import LExperimentDetails from '@/components/molecules/LExperimentDetails.vue'
-import LExperimentsDrawer from '@/components/molecules/LExperimentsDrawer.vue'
-import LExperimentResults from '@/components/organisms/LExperimentResults.vue'
-import LJobResults from '@/components/molecules/LJobResults.vue'
-import LExperimentLogs from '@/components/molecules/LExperimentLogs.vue'
-import LExperimentsEmpty from '@/components/molecules/LExperimentsEmpty.vue'
+import { useSlidePanel } from '@/composables/useSlidePanel'
+import LPageHeader from '@/components/layout/LPageHeader.vue'
+import LExperimentTable from '@/components/experiments/LExperimentTable.vue'
+import LExperimentForm from '@/components/experiments/LExperimentForm.vue'
+import LExperimentDetails from '@/components/experiments/LExperimentDetails.vue'
+import LExperimentsDrawer from '@/components/experiments/LExperimentsDrawer.vue'
+import LExperimentResults from '@/components/experiments/LExperimentResults.vue'
+import LJobResults from '@/components/experiments/LJobResults.vue'
+import LExperimentLogs from '@/components/experiments/LExperimentLogs.vue'
+import LExperimentsEmpty from '@/components/experiments/LExperimentsEmpty.vue'
 import type { Experiment, Job } from '@/types/Experiment'
 
 const { showSlidingPanel } = useSlidePanel()
