+++ lumigator/frontend/src/composables/useSlidePanel.ts
